fix(e2e): Linux build (IPPROTO_TCP Int32) + Samba postinst (smb.conf を install 後 cp)
- POSIXSocketTransport: glibc の IPPROTO_TCP は Int なので Int32() に包む (CI build 検出)
- e2e.yml: smb.conf を /etc/samba に :ro bind-mount すると samba-common postinst が
  read-only ファイルに書けず install 失敗 → /tmp に mount し install 後に cp する

+++ Sources/SMBee/POSIXSocketTransport.swift
@@ -64,7 +64,8 @@ public final class POSIXSocketTransport: SMBTransport, @unchecked Sendable {
         #else
         hints.ai_socktype = SOCK_STREAM
         #endif
-        hints.ai_protocol = IPPROTO_TCP
+        // glibc では `IPPROTO_TCP` が `Int`、Darwin では `Int32`。両対応で Int32 に包む。
+        hints.ai_protocol = Int32(IPPROTO_TCP)
         var result: UnsafeMutablePointer<addrinfo>?
         let service = String(port)
         guard getaddrinfo(host, service, &hints, &result) == 0, let first = result else {
